Use GSettings instead of GConf
GConf has been deprecated for some time now. This change removes the
ability to set a local musicbrainz server. The support for this was
half-hearted as it was only possible to set a custom hostname, not a
custom port. This functionality could be restored fairly easily if
there is a demand for it.

Saving the window size and state is also removed as this should be
stored in the xdg user config dir instead. This will be implemented
soon.

Thanks to Gert Kulyk & Phillip Wood for their contributions.

https://bugzilla.gnome.org/show_bug.cgi?id=625897