Fix device names on Solaris (#540857)
Solaris is unique that it separates /dev/dsk and /dev/rdsk where /dev/dsk is
the logical name for the drive and /dev/rdsk is the raw device name.  You need
to use the raw device name to acutally access the device.  For example, you can
refer here for more information:

  http://docs.sun.com/app/docs/doc/801-7038/6i1cfovsu?a=view

Note that the sound-juicer code already has some code to manage this in the
function sj_metadata_getter_set_cdrom() in libjuicer/sj-metadata-getter.c.

The attached patch adds 3 more places in the code where "/dev/rdsk" is needed
for sound-juicer to work well on Solaris.  It also fixes the one place in
libjuicer/sj-metadata-getter.c which already references "/dev/rdsk" so the
 #ifdef matches the others.  THere is no need to also check for "__SVR4".

Signed-off-by: Ross Burton <ross@linux.intel.com>