Use relationlist_list_foreach_relation in fill_relations
Now that we have a helper to iterate over all relations, better
to use it as much as possible. It would probably be more efficient
to have a relationlist_list_foreach helper iterating over RelationList
items contained in a RelationListList, and a relationlist_foreach helper
iterating over Relation items contained in a RelationList.
This would allow fill_relations to avoid iterating over and over
on the RelationListList while it's always only interested in the
'url' RelationLists.

https://bugzilla.gnome.org/show_bug.cgi?id=661646