Convert to single column.
At the moment the user inevitably has to scroll to the right to see
the artist when choosing between multiple releases. As there are only
ever a handful of releases to choose between it would be better to
display the artist below album title so all the information is visible
at once in a single column. This will be especially true once we add
the release label, year and country as well. Make the album title bold
and the artist name italic to visually differentiate between them.

https://bugzilla.gnome.org/show_bug.cgi?id=674926