Zero-pad track number in %dn pattern
59fc2c6523f changed the "Number - Title" file pattern to use %dn
instead of %tN. However, the track number is not 0-padded in %dn,
which gives badly sorted tracks in some situations, and was
the cause of several bug reports. This commit changes the %dn pattern
to 0-pad the track number to restore things as they were when %tN was
used. Let's hope there were not many users who preferred non-zero-padded
filenames...

https://bugzilla.gnome.org/show_bug.cgi?id=677865