Don't try to free an uninitialized pointer
brasero_drive_lock() is documented as filling a 'reason_for_failure'
arg with a description string when an error occurs. However, the
method does nothing with this parameter, it does not even initialize
it. This causes sound-juicer to crash when hitting this code path
as it's assuming this varialble will be set in error cases,
and it's trying to free this uninitialized pointer.
Avoid this crash by initializing 'reason' to NULL, g_free() will
be a no-op..

https://bugzilla.gnome.org/show_bug.cgi?id=665881