Fix play seek scale.
In the transition from glade to gtkbuilder the adjustment for
seek_scale went AWOL. As a result that slider didn't move in playback
and trying to seek with the mouse rewound the track to the start
because the range for the scale wasn't properly initialized. This fix
adds the adjustment back into the UI file, updates the scale widget to
GtkScale and fixes seeking with the keyboard as well.

N.B. The keyboard fix bumps the minimum GTK version requirement to 3.2

https://bugzilla.gnome.org/show_bug.cgi?id=637867