Remove remaining stock items
This removes the arrow on the 'Continue' button in the multiple album
dialog partly to conform with the recommendation that buttons just
have a label describing their action and partly to avoid having to
deal with the RTL variant of the icon. The Alt+N accelerator on
'Cancel' in the infobar is non standard but avoids a clash with the
album composer entry. Alt+P for Pause is non standard but avoids a
clash with the album artist entry and gives a consistent accelerator
for Play/Pause. The disc number entry was missing an accelerator so it
is assigned Alt+D.

https://bugzilla.gnome.org/show_bug.cgi?id=705131