Use localized artist aliases from MusicBrainz.
MusicBrainz returns artist names in their native locale so it returns
"Пётр Ильич Чайковский" instead of "Pyotr Ilyich Tchaikovsky" but
provides localized versions of the artist name as aliases. Use the
alias so we get "Pyotr Ilyich Tchaikovsky" in English locales and
"Piotr Ilitch Tchaïkovski" in French locales etc.

Although GLib supports fallback locales we do not use them as
MusicBrainz does not provide locale information for the canonical
artist name, only it's aliases. This means that it is not possible to
fallback to other locales as the canonical name may actually be in a
higher priority locale than the alias that matches one of the fallback
locales.

https://bugzilla.gnome.org/show_bug.cgi?id=723297