Add SjCellRendererText
This is a subclass of GtkCellRendererText which accepts changes
without confirmation and does not stop editing when the toplevel
loses focus. This means that the user no longer loses their changes
if they forget to press 'Enter' when they have finished editing.

This raises the required GLib version to 2.38 for
G_DEFINE_TYPE_WITH_PRIVATE

https://bugzilla.gnome.org/show_bug.cgi?id=151469