Add SjTreeView
This is a subclass of GtkTreeView where horizontal movements skip
columns that are not activatable/editable. Tab moves to the next
activatable/editable wrapping if necessary. Control-Tab moves the
keyboard focus to the next widget. With a suitable cell renderer
moving cells while editing is also possible so the user does not need
to press 'Enter' to confirm the changes, then move and then press
'Enter' again to start editing the new cell.

https://bugzilla.gnome.org/show_bug.cgi?id=151469