Add movement keys to SjCellRendererText
Allow the user to move cells while editing. This makes editing much
quicker as the user no longer has to press 'Enter' to confirm their
changes, then move and then press 'Enter' again to start editing the
new cell.

[Shift]Tab - move to the [previous]next activatable/editable column
[Page]Up - move up a [page]row
[Page]Down - move down a [page]row
Control-Home - move to the top of the current column
Control-End - move to the bottom if the current column
Control-Tab - move the keyboard focus to the next widget

https://bugzilla.gnome.org/show_bug.cgi?id=151469