libjuicer: Fix empty fallback titles with some CDs
My CD of "A Movie for Daddy" has some CD-Text information, but
simply empty ones. Tracks 4 and 10 have " " as track titles, which
we should ignore.

Rework the setting of metadata to ignore strings that are full of
spaces. This also fixes a small memory leak if the string contained
a '\0' instead of simply being NULL.

https://bugzilla.gnome.org/show_bug.cgi?id=729916