Remove duplicate recording labels
If a release has more than one catalog number on the same recording
label then the label name is duplicated as it appears with each
catalog number. As we're not using the catalog numbers skip adding the
recording label to the list if it's already present.

For an example album see
https://musicbrainz.org/ws/2/release/082bc530-580c-4628-8aa9-9003118ffa8b?inc=labels