Edit text of multiple tracks at once
This is the final step in allowing the user to select and edit
multiple rows at once. Implement editing for text fields of all
selected rows and activate multiple selections on the track
listview. This makes it much quicker to edit a common Artist/Composer
for a subset of tracks. As it is possible to activate (and therefore
start editing) a cell that is not in the selection take care to only
update the selected rows if the activated cell is selected.

https://bugzilla.gnome.org/show_bug.cgi?id=151469