Edit extract column of multiple tracks at once
Allowing the user to select multiple rows at once makes it much
quicker to edit common fields. Prepare the way by implementing this
for the extract column. As it is possible to activate a cell that is
not in the selection take care to only update the selected rows if the
activated cell is selected.

https://bugzilla.gnome.org/show_bug.cgi?id=151469