Fix setting of Music folder in preferences dialog
Selecting a folder from the drop-down list wasn't changing the folder
that sound-juicer used to save extracted tracks. This is because
'current-folder-changed' only fires if the user changes the folder in
the file chooser dialog. It does not fire if the user selects a folder
from the drop-down list provided by the file chooser button. This
caused sound-juicer to ignore the change if a user selected a folder
from the list. Fix this by connecting to 'file-set' instead.