Add SjTreeView::play-cursor-row and SjTreeView::play-row
As SjTreeView::move-cursor restricts keyboard cursor movements to
activatable/editable columns it's no longer possible to play the track
under the treeview cursor with the keyboard. Add two signals to fix
this: 'play-row' for playing the specified path and 'play-cursor-row'
for playing the cursor row. 'play-cursor-row' is bound to
Control-Space.

https://bugzilla.gnome.org/show_bug.cgi?id=151469