Don't retrieve metadata twice at start up
When sound-juicer starts up it calls reread_cb() twice, once from
set_device() and once from media_added_cb(). Add a check to
media_added_cb() to see if the metadata is already being retrieved
before calling reread_cd(). If the user changes the device while
retrieving metadata we still start a new request for the new drive. It
would be nice to make metadata retrieval cancellable in the future.

https://bugzilla.gnome.org/show_bug.cgi?id=705159