Move submit infobar to ui file
The infobar at is only ever used to prompt the user to submit an album
to Musicbrainz so there is no need to keep creating and destroying
it. Moving it definition to the ui file simplifies the code and
removes code from GEdit which does not have an exception for
proprietary codecs.

The appearance of the infobar is changed slightly, there is no icon
anymore and it now uses a symbolic close button which saves some
space.

Note that the buttons in the infobar are packed into their own box to
stop the button box adding internal padding to the buttons which makes
the close button look too wide.

This raises the required Gtk version to 3.4 for GtkActionable