Use non recursive make
Non recursive make is more robust and efficient. It will also make it
simpler to use vala in the future as we wont need to create a separate
vapi file for the code in each subdirectory.

Libtool is no longer used for the build (libjuicer needs some work to
use opaque data structures and have its own translation domain before it
can be installed as a shared library). Not using libtool means that the
object file dependencies of tests need to be manually added but has the
benefit that are only rebuilt if their specific dependencies have
changed which means one can lazily rerun only the updated tests with

env RECHECK_LOGS= make -e check

sound-juicer is now built in the root directory, data file paths have
been updated to allow uninstalled running from that directory.

https://bugzilla.gnome.org/show_bug.cgi?id=741288