Fix crash on FreeBSD
Passing a NULL string to mb5_query_set_proxyhost(),
mb5_query_set_proxyusername() or mb5_query_set_proxypassword() causes a
crash on FreeBSD so pass an empty string instead. This is caused by
libmusicbrainz passing the NULL string to the constructor of std::string
which is undefined behaviour in the C++ standard.

https://bugzilla.gnome.org/show_bug.cgi?id=742019