Improve eject responsiveness
It takes a couple of seconds from pressing the eject keyboard shortcut
for the UI to update as there is a lag while the drive opens. To fix
this update the UI when the shortcut is pressed and check in
media_removed_cb() to see whether it's already been updated or not.

https://bugzilla.gnome.org/show_bug.cgi?id=753509