Improve wording of submit infobar label
Try to make it clearer that the user should edit the track names in
MusicBrainz if they want to submit a new album. Currently users are
confused when they edit the track names in Sound Juicer and it does not
submit their edits.

Change the wording of the menu item for submitting track names to make
it clearer as well.

https://bugzilla.gnome.org/show_bug.cgi?id=753330