Zero pad disc number when needed
Zero pad the disc number appropriately for the number of discs in the
current album. This means that the tracks will sort properly when listed
alphabetically by name. Currently there is no way to enter the number of
discs in the current album in the UI so this only works for albums that
have that information in MusicBrainz.

https://bugzilla.gnome.org/show_bug.cgi?id=753739