Add barcode and catalog number to multiple_album_dialog
Extend the details on the possible albums shown by the
multiple_album_dialog by the barcode and the catalog number used by
the label releasing the album/release.  This eases the selection of
the correct musicbrainz release, when an album was released several
times and the release label and country do not offer enough
information to identify the release.

https://bugzilla.gnome.org/show_bug.cgi?id=757156