Limit range of EggPlayPreview scale
When moving the scale with the mouse if one moves past the end of the
scale the value seen by the ‘change-value’ handler keeps changing
which means that time label could show times greater than the track
length. Fix this by checking that the value is within the permitted
range.

Note that handling ‘change-value’ is preferable to handling
‘value-changed’ here as ‘change-value’ is only emitted in response to
scroll events so it avoids a feedback loop which would exist between a
‘value-changed’ handler and timeout_cb() which calls
gtk_range_set_value().