Improve accuracy of EggPlayPreview seek scale
The seek scale would sometimes move backwards due to rounding errors
and would jump on short tracks due to the relatively long update
interval. Fix by using GstClockTime for the position and duration and
increasing the frequency of the callback that updates the scale
position. Also simplify the logic for checking if we're currently
seeking.