Retry query if MusicBrainz is busy
Recently MusicBrainz has been returning error 503 for some queries when
it is overloaded. This causes the lookup to fail. If it’s busy wait a
while and retry the query. To avoid overloading MusicBrainz even more we
limit the number of retries to 4 and wait 4 seconds for the first retry,
a further 11 seconds for the second and then 30 seconds between
subsequent retries.

https://bugzilla.gnome.org/show_bug.cgi?id=764940

https://bugzilla.gnome.org/show_bug.cgi?id=764941