Check MCN before querying detailed album metadata
Currently Sound Juicer queries MusicBrainz for detailed album metadata
for all the releases that match the disc id in turn, stopping if it
finds one where the barcode matches the MCN on the disc. This means that
unless the barcode for the first release that is queried matches the MCN
it is performing unnecessary queries which slows down the metadata
retrieval.

The query that returns the matching album id’s for the disc id also
returns the barcodes so it is possible to check the barcodes against the
MCN and if one matches only perform the detailed release query for that
release. This speeds up getting the metadata for discs with multiple
matches where the MCN matches the barcode of one of the releases.

https://bugzilla.gnome.org/show_bug.cgi?id=764941