Stop translating icon name
Gettext automatically extracts the icon name from the desktop file for
translation¹ because it is a locale string in the desktop entry
specification.² However we don’t want it translated. As it’s not
possible to specify per-language keywords to stop xgettext from
extracting the string use sed to remove it from the pot file.

Note that we need to leave the translator comment about the icon name
in the desktop file for users of damned-lies as it does not run make
to extract the strings and ends up extracting the icon name. The changes
to the makefile rule to build the translated desktop file should protect
us from any accidental translations.

¹ https://savannah.gnu.org/support/?108887
² http://standards.freedesktop.org/desktop-entry-spec/latest/ar01s05.html

https://bugzilla.gnome.org/show_bug.cgi?id=769623