Stop translating icon name
Gettext automatically extracts the icon name from the desktop file for
translation¹ because it is a locale string in the desktop entry
specification.² However we don’t want it translated. As it’s not
possible to specify per-language keywords to stop xgettext from
extracting the string use sed to remove it from the pot file.

¹ https://savannah.gnu.org/support/?108887
² http://standards.freedesktop.org/desktop-entry-spec/latest/ar01s05.html

https://bugzilla.gnome.org/show_bug.cgi?id=769623