Add SjAlbumChooserDialog
A template widget replacement for the current multiple album dialog
implementation with the following improvements:

 • Respect GtkSettings:gtk-dialogs-use-header.
 • Use the new GtkScrolledWindow:max-content-height/width properties to
   set an appropriate window size for the number of rows being displayed.†
 • Use GtkListBox rather than GtkTreeView.

This commit adds the new code for the dialog - it does not populate the
rows yet, the next commit with copy and paste the code for that from
sj-main.c.

This commit raises the required Gtk version to 3.21.3.

† For this to work a new dialog instance must be created for each new
  album list.