Add some translator comments to desktop file
Gettext automatically extracts the icon name from the desktop file for
translation¹ because it is a locale string in the desktop entry
specification.² However we don’t want it translated so add a comment to
this effect. Also add a comment that the translated keyword list must
contain semicolons.

Thanks to Piotr Drąg for suggesting these changes.

¹ https://savannah.gnu.org/support/?108887
² http://standards.freedesktop.org/desktop-entry-spec/latest/ar01s05.html

https://bugzilla.gnome.org/show_bug.cgi?id=769623