Use GtkScrolledWindow:propagate-natural-height/width
GtkScrolledWindow has reverted to returning the minimum size as the
natural size as well. This makes the dialog too small. Use the new
properties so the dialog is sized on the natural size of the listbox
(clamped by the max-context-height/width properties).

NB This requires Gtk from git master as the
propagate-natural-height/width properties are not in a Gtk development
release.