MusicBrainz: Use Fuzzy TOC matching
When searching for the disc on MusicBrainz use fuzzy TOC matching so
that if the discid isn’t found but there is a album with the same
track lengths that is returned. Disable cdstub matches as we cannot
parse them at the moment. The fuzzy match will not have a matching
discid so match track lengths against the disc we’re ripping to find
out which disc in a release matches our disc.

https://bugzilla.gnome.org/show_bug.cgi?id=788425